Move the User Interface module under Runtime
With this, Boss will again only coordinate other modules.
