+++ configure.ac
@@ -153,7 +153,6 @@ AC_CONFIG_FILES([Makefile
                  pyanaconda/modules/boss/install_manager/Makefile
                  pyanaconda/modules/boss/kickstart_manager/Makefile
                  pyanaconda/modules/boss/module_manager/Makefile
-                 pyanaconda/modules/boss/user_interface/Makefile
                  pyanaconda/modules/security/Makefile
                  pyanaconda/modules/timezone/Makefile
                  pyanaconda/modules/network/Makefile
@@ -185,6 +184,7 @@ AC_CONFIG_FILES([Makefile
                  pyanaconda/modules/payloads/source/url/Makefile
                  pyanaconda/modules/runtime/Makefile
                  pyanaconda/modules/runtime/dracut_commands/Makefile
+                 pyanaconda/modules/runtime/user_interface/Makefile
                  pyanaconda/modules/storage/Makefile
                  pyanaconda/modules/storage/bootloader/Makefile
                  pyanaconda/modules/storage/checker/Makefile
