+++ pyanaconda/input_checking.py
@@ -26,7 +26,7 @@ from pyanaconda.core import constants, regexes
 from pyanaconda.core import users
 from pyanaconda.anaconda_loggers import get_module_logger
 from pyanaconda.modules.common.constants.objects import USER_INTERFACE
-from pyanaconda.modules.common.constants.services import BOSS
+from pyanaconda.modules.common.constants.services import RUNTIME
 from pyanaconda.modules.common.structures.policy import PasswordPolicy
 
 log = get_module_logger(__name__)
@@ -38,7 +38,7 @@ def get_policy(policy_name) -> PasswordPolicy:
     :param policy_name: a name of the policy
     :return: a password policy data
     """
-    proxy = BOSS.get_proxy(USER_INTERFACE)
+    proxy = RUNTIME.get_proxy(USER_INTERFACE)
     policies = PasswordPolicy.from_structure_dict(proxy.PasswordPolicies)
 
     if policy_name in policies:
